// client/src/interface/Availability.ts
export interface Availability {
  monday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  };
  tuesday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  };
  wednesday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  };
  thursday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  };
  friday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  };
  saturday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  };
  sunday: {
    isAvailable: boolean;
    time: {
      from?: string;
      to?: string;
    };
  }
}